requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiohttp-cors==0.8.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
beartype==0.15.0
bleach==6.2.0
bokeh==3.7.3
boto3==1.35.99
botocore==1.35.99
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
contourpy==1.3.2
dataclasses-json==0.6.7
decorator==5.2.1
deltalake==0.17.4
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
executing==2.2.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
frozenlist==1.6.0
fs==2.4.16
fsspec==2025.5.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.3
google-cloud-pubsub==2.29.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.31.0
gradio_client==1.10.1
greenlet==3.2.2
groovy==0.1.2
groq==0.25.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h3==4.2.2
hf-xet==1.1.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.0
idna==3.10
importlib_metadata==8.6.1
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.15
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.42
linkify-it-py==2.0.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.40.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.2.2
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
panel==1.7.0
param==2.2.0
parso==0.8.4
pathway==0.21.5
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
plotly==6.1.1
plotly-express==0.4.1
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-sat==1.8.dev16
pytz==2025.2
pyviz_comms==3.0.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
ruff==0.11.11
s3transfer==0.10.4
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
setuptools==80.8.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlglot==10.6.1
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
tomlkit==0.13.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
tweepy==4.15.0
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xyzservices==2025.4.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
